--- a/go.mod
+++ b/go.mod
@@ -3,11 +3,15 @@ module github.com/tarantool/go-tarantool
 go 1.11
 
 require (
+	github.com/google/go-cmp v0.5.7 // indirect
 	github.com/google/uuid v1.3.0
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/shopspring/decimal v1.3.1
 	github.com/stretchr/testify v1.7.1 // indirect
 	github.com/tarantool/go-openssl v0.0.8-0.20220419150948-be4921aa2f87
 	google.golang.org/appengine v1.6.7 // indirect
 	gopkg.in/check.v1 v1.0.0-20200227125254-8fa46927fb4f // indirect
 	gopkg.in/vmihailenco/msgpack.v2 v2.9.2
+	gotest.tools/v3 v3.2.0 // indirect
 )