See also:
- https://github.com/msgpack/msgpack/blob/master/spec.md
- https://www.tarantool.io/en/doc/latest/dev_guide/internals/msgpack_extensions/#the-decimal-type

Follows up https://github.com/tarantool/tarantool/issues/692
Closes #96