@ligurio 2022-05-12
New changes:

- parametrize e2e test (`TestInsert`) with test samples
- enable fuzzing in CI
- add constant with max decimal precision and use it in tests and package itself