@funny-falcon 2022-05-15
И ещё один коммит пушнул. Попробовал проверить количество цифр.
И думаю, что это не последняя проверка, которая нужна. Проверь пример из последнего коммита.