@ligurio 2022-05-17
> В общем, я пушну свой вариант bcd.go .

@funny-falcon  Don't touch my branch. I made a separate branch for you - `funny-falcon/decimal-playground`. Feel free to experiment there.