@ligurio 2022-05-23
Rewrote a BCD decoding function (`DecodeStringFromBCD()`) to reduce a number of allocations and slices copying.
Benchmarks report significant improvements:
```
name                                    old time/op    new time/op    delta
MPEncodeDecodeDecimal/0.3-8               2.73µs ± 2%    2.41µs ± 3%  -11.92%  (p=0.000 n=8+8)
MPEncodeDecodeDecimal/-18.34-8            3.26µs ± 2%    2.78µs ± 1%  -14.68%  (p=0.001 n=7+7)
MPEncodeDecodeDecimal/-108.123456789-8    3.82µs ± 4%    3.04µs ± 2%  -20.30%  (p=0.000 n=8+8)
...
DecodeStringFromBCD/0.3-8                  278ns ± 0%      48ns ± 0%  -82.63%  (p=0.000 n=8+8)
DecodeStringFromBCD/-18.34-8               462ns ± 3%      60ns ± 2%  -86.92%  (p=0.000 n=8+8)
DecodeStringFromBCD/-108.123456789-8       722ns ± 9%      89ns ± 2%  -87.68%  (p=0.000 n=8+8)

name                                    old alloc/op   new alloc/op   delta
MPEncodeDecodeDecimal/0.3-8               1.39kB ± 0%    1.26kB ± 0%   -9.20%  (p=0.000 n=8+8)
MPEncodeDecodeDecimal/-18.34-8            1.74kB ± 0%    1.38kB ± 0%  -20.28%  (p=0.000 n=8+8)
MPEncodeDecodeDecimal/-108.123456789-8    2.27kB ± 0%    1.54kB ± 0%  -32.39%  (p=0.000 n=8+8)
...
DecodeStringFromBCD/0.3-8                   131B ± 0%        3B ± 0%  -97.71%  (p=0.000 n=8+8)
DecodeStringFromBCD/-18.34-8                360B ± 0%        8B ± 0%  -97.78%  (p=0.000 n=8+8)
DecodeStringFromBCD/-108.123456789-8        752B ± 0%       16B ± 0%  -97.87%  (p=0.000 n=8+8)

name                                    old allocs/op  new allocs/op  delta
MPEncodeDecodeDecimal/0.3-8                 51.0 ± 0%      47.0 ± 0%   -7.84%  (p=0.000 n=8+8)
MPEncodeDecodeDecimal/-18.34-8              57.0 ± 0%      52.0 ± 0%   -8.77%  (p=0.000 n=8+8)
MPEncodeDecodeDecimal/-108.123456789-8      60.0 ± 0%      54.0 ± 0%  -10.00%  (p=0.000 n=8+8)
...
DecodeStringFromBCD/0.3-8                   5.00 ± 0%      1.00 ± 0%  -80.00%  (p=0.000 n=8+8)
DecodeStringFromBCD/-18.34-8                6.00 ± 0%      1.00 ± 0%  -83.33%  (p=0.000 n=8+8)
DecodeStringFromBCD/-108.123456789-8        7.00 ± 0%      1.00 ± 0%  -85.71%  (p=0.000 n=8+8)
```

@0x501D, could you take a look to changes (in a separate commit in the branch)?