@funny-falcon 2022-05-23
Ну вот, другое дело.

Теперь ещё EncodeStringToBCD переписать, и будет хорошо.

И isNegative - можно использовать массив:
```
var isNegative = [256]bool{
	0x0a: false,
	0x0b: true,
	0x0c: false,
	0x0d: true,
	0x0e: false,
	0x0f: false,
}
```