@oleg-jukovec 2022-06-22
@Totktonada said I could continue work in the pull request. Is it okay?