@oleg-jukovec 2022-06-22
I made a lot of changes without consulting with @ligurio , so I added myself as co-author of this commits. This is done so that @ligurio will not be shamed for my changes. I had no intention of appropriating his great work.